fixed loops for checking to send to user

+++ modules/info.py
@@ -52,6 +52,8 @@ class Information(commands.Cog):
                 userid = person['userid']
                 added = person['added']
                 user = self.bot.get_user(userid)
+                if user is None:
+                    continue
                 if added:
                     await self.bot.sendSuccess(user, f"You have had one premium credit added to your account! Use the `redeem` command to get started!")
                 else:
@@ -69,10 +71,12 @@ class Information(commands.Cog):
                 userid = person['userid']
                 receiveCredit = person['receivecredit']
                 user = self.bot.get_user(userid)
+                if user is None:
+                    continue
                 if receiveCredit:
                     await self.bot.sendSuccess(user, f"You have had one premium credit added to your account! Use the `redeem` command to get started! Thank you for voting for Ticketer!")
                 else:
-                    await self.bot.sendSuccess(user, f"Thank you for voting for Ticketer! You currently have {cur_votes} votes on your account Continue voting to receive 1 Premium Credit.")
+                    await self.bot.sendSuccess(user, f"Thank you for voting for Ticketer! You currently have **{cur_votes} votes** on your account Continue voting to receive 1 Premium Credit.")
                 await self.bot.db.execute("DELETE FROM votesqueue WHERE userid = $1", userid)
             await asyncio.sleep(30)
